# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "First_proj",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a list of restaurants",
                "operationId": "2f995b8db474e63c60fccf4e231ff654",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "NUmber of elements on page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Asador Etxebarri"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Filter by country",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "France"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Asador Etxebarri"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "France"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 50.42297
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 30.24416
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-08-15T08:55:40.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-08-15T08:55:40.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagintaion": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 23
                                                },
                                                "perPage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "lastPage": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create new Restaurant",
                "description": "Returns Restaurant data",
                "operationId": "createRestaurant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Diverxo"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50.59067
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30.71773
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Diverxo"
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "France"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 50.59067
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 30.71773
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-08-21T13:40:26.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-08-21T13:40:26.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 22
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/restaurants/delete/{id}": {
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete the Restaurant",
                "description": "Returns message about operation",
                "operationId": "deleteRestaurant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Restaurant's id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant was removed"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}